chore(gui): aligned text copying location with it's common usage
+  aligned text copying with updated design
