+++ frontend/src/js/components/settings/user-management/selfusermanagement.js
@@ -23,13 +23,13 @@ import { getCurrentSession, getCurrentUser, getFeatures, getIsDarkMode, getIsEnt
 import { editUser, saveUserSettings } from '@northern.tech/store/thunks';
 
 import { toggle } from '../../../helpers';
+import { CopyTextToClipboard } from '../../common/copytext';
 import ExpandableAttribute from '../../common/expandable-attribute';
 import Form from '../../common/forms/form';
 import PasswordInput from '../../common/forms/passwordinput';
 import TextInput from '../../common/forms/textinput';
 import InfoText from '../../common/infotext';
 import AccessTokenManagement from '../accesstokenmanagement';
-import { CopyTextToClipboard } from '../organization/organization';
 import TwoFactorAuthSetup from './twofactorauthsetup';
 import { UserId, getUserSSOState } from './userdefinition';
